Fix NodeDispatchGrid diagnostic check
The product of the components of the NodeDispatchGrid or
NodeMaxDispatchGrid attributes may not exceed 16,777,215 (2^24-1)
but this was not diagnosed correctly in cases where the product
overflowed 32 bits.
Fixed by increasing the calculation of the product to 64 bits.

+++ tools/clang/lib/Sema/SemaHLSL.cpp
@@ -13026,7 +13026,7 @@ void Sema::DiagnoseGloballyCoherentMismatch(const Expr *SrcExpr,
 
 void ValidateDispatchGridValues(DiagnosticsEngine &Diags,
                                 const AttributeList &A, Attr *declAttr) {
-  unsigned x = 1, y = 1, z = 1;
+  uint64_t x = 1, y = 1, z = 1;
   if (HLSLNodeDispatchGridAttr *pA =
           dyn_cast<HLSLNodeDispatchGridAttr>(declAttr)) {
     x = pA->getX();
